Move generic app runtime into app module
